Added persistence helpers to the jobs service
ref https://linear.app/ghost/issue/ENG-2027/

The email analytics job has had a one-off/custom implementation of using the jobs table to persist `started_at` and `finished_at` timestamps for the analytics jobs. This would be helpful for other scheduled/recurring jobs, so this adds helper fns to the JobsService to do so.

The intent is for an entry in the `jobs` table to simply hold the last started/finished times so we don't need to schedule a random time on Ghost boot. This ensures jobs are run daily, etc, in the event of frequent reboots.