🐛 Fixed subscription attribution sometimes missing from `SubscriptionCreatedEvent` (#22219)
ref https://linear.app/ghost/issue/ENG-1561/

Depending on the path to upgrade a member, it was possible for the subscription attribution to not get set. This specifically patches the `SubscriptionCreatedEvent` missing the subscription attribution.

Note that we've identified other issues while exploring this causing subscription attribution to not display in certain cases, although it is being set appropriately. Those will be patched separately. See parent issue.