Provided attribution directly to SubscriptionActivatedEvent for upgrading members
ref https://linear.app/ghost/issue/ENG-2030/

We could experience a race condition on upgraded members where the SubscriptionActivatedEvent would look to the SubscriptionCreatedEvent for subscription attribution, but since both are emitted at the same time, it was possible for the Created event to not exist.

This commit passes the attribution directly to the SubscriptionActivatedEvent when creating both events in `linkSubscription`, thereby avoiding the race condition.

We maintain the fallback behavior to looking to the SubscriptionCreatedEvent in the StaffService email code to cover the other flows.