Integrating React Router in ActivityPub (#22234)
ref AP-701

- ATM we're using our in-house routing solution in ActivityPub. On the
mid/long term (ie. Public Beta / GA) it's beneficial to switch to an
existing third-party solution which handles certain use cases that we'd
need out of the box. One of the most popular of these tools is React
Router, for which we've already put together a spike in the Posts app
prototype. This PR is replacing our in-house router with React Router
for ActivityPub and all new React apps.