Prevent sendAuthCodeToUser from throwing 500s
ref ENG-2004

There are intermittent 500s from this codepath, seemingly because the
user id isn't always valid. I've tested with cookies disabled, and a
bunch of different flows but haven't been able to repro the issue. For
now the best path seems to be handling the error better, so we get a
better stack trace.