Added api versions dropdown to stats page
ref https://linear.app/ghost/issue/ANAL-123/

This is intended to help with iterating on tinybird API testing. On the `/stats/` page we now have a dropdown for the API version that applies to all fetched endpoints. This was easier than hard-coding to config, which we aren't even certain is the correct approach here.

The default is v0 which is the current behavior.