Added custom analytics tracking script
ref https://linear.app/ghost/issue/ANAL-126/
- moved away from flock.js script while we wait for PRs to be merged
- added new tracker in publicly-served files
- added datasource config so we can pivot without changes to Ghost code

We needed to move away from the flock.js script because it doesn't allow the JSON data type for testing the front-end script with the tinybird instance that supports the JSON data type.

There's a PR (https://github.com/tinybirdco/web-analytics-starter-kit/pull/129) to add this configurability to their script so we can revert these changes.

Added rendering tests for ghost head for including the script. Previously we only tested that the head would render.