Improved branch naming for Tinybird test job (#22339)
ref
https://linear.app/ghost/issue/ANAL-137/experiment-develop-an-mvp-of-our-cicd-workflows

- The initial implementation of this job creates a TB branch using the
default heuristic provided by Tinybird, and results in some slightly
weird/imperfect branch names.
- In particular, the `_NORMALIZED_BRANCH_NAME` evaluates to an empty
string, so there's a weird `__` in the branch name. More importantly,
when the workflow runs on `main`, there is no PR number so the result is
just `tmp_ci__` which won't work if we have multiple commits merged to
main, running this job simultaneously.
- This commit updates the logic for the branch name to be either
`tmp_ci_pr_{PR number}` if the workflow was triggered by PR, or
`tmp_ci_main_{first 7 chars of commit sha}` if triggered by a push to
main.
- It also DRY's up the logic for generating the branch name, so we only
have to change it in one place.