Improved error message if secrets are missing in CI run (#22342)
ref
https://linear.app/ghost/issue/ANAL-137/experiment-develop-an-mvp-of-our-cicd-workflows

- Github Actions doesn't allow repository secrets to be accessed in CI
runs when the PR is opened from a fork. This is an intentional design
"feature" for security purposes, so there aren't any legitimate /
risk-free workarounds.
- This results in some frustratingly vague error messages if you open a
PR from a fork and it fails, usually in the form of some "authentication
failed" message, which doesn't immediately point to a solution.
- This commit adds an explicit check for access to the required secrets
for both the browser tests and the tinybird tests (the only two jobs
that I'm aware of that depend on secrets being available), and outputs a
user-friendly error message if it fails, directing the user to re-open
the PR from the upstream TryGhost/Ghost repo