Updated tinybird test job to always delete the test branch (#22373)
no issue

- If the `job_tinybird` test failed, it would leave a tinybird branch
behind because it didn't make it to the "delete branch" step. With a max
of 3 branches allowed in Tinybird, it only took a few of these instances
to meet that limit, and all other CI runs would then fail.
- This ensures the "delete branch" step runs even if the tests pass, so
we shouldn't ever run out of available branches, unless we're actually
running >3 tests simultaneously