Modified tinybird job to deploy to staging and production (#22393)
no issue

- We've been using a single workspace in Tinybird for both staging and
production, and we want to start using a separate workspace for each
environment.
- We've also have two Tinybird directories - `ghost/tinybird` and
`ghost/tinybird-dedicated`, and we're planning to clean both of those up
and start fresh in `ghost/web-analytics`.
- This PR keeps the Tinybird tests running against our staging
workspace, and modifies the deploy job to deploy to both staging and
production when merging to main (at least for now)