Improved text for email MFA requests (#22545)
ref ENG-2074

The email text can be changed based on whether the verification code is
available, but the messages in Ghost Admin can't as they rely on the
settings API (which is authenticated). Instead they're simplified to
remove the content about an unrecognised device, which could be
incorrect if 2FA is required for every login.