Added naive stats button to Post analytics (#22555)
ref ANAL-141

- Traffic analytics can be filtered for pages but only from within
stats. This PR adds a button in the Post analytics which allows direct
filtering of the post in traffic analytics. This is a naive
implementation as it simply redirects to Stats and prefilters for the
slug, it doesn't check if any data exists for the post in analytics.
Also since it's based on slug and not post.id it only works until the
URL/slug is not changed.