Fixed mismatch in visits and unique visits on stats page (#22588)
ref https://linear.app/ghost/issue/ANAL-96
- fixed mismatch in visits and unique visits
- moved away from sessions-based materialized views (still hanging
around)
- bounce rate needs revisions

This doesn't quite fix everything, as bounce rate looks off yet. It's
much better than it was and I'll follow up tomorrow/next week with the
rest.

We also need to clean up or consolidate the views, which I need to think
about more before getting rid of what we have.