Fixed Tinybird timezone bug (#22656)
ref
https://linear.app/ghost/issue/ANAL-175/stats-page-doesnt-update-data-if-youre-behind-utc-and-its-already

There was a bug in our Tinybird pipes that ignored the site's timezone
setting. This could cause it to _appear_ that the site wasn't ingesting
data anymore, even though it actually was.

This commit adds a `timezone` query parameter to all the pipes used in
Admin and converts the timestamps to this timezone before
aggregating/filtering the dataset.

It also adds the stats_page token to all pipes, and bumps the TB_VERSION
in admin to 7.