Switched traffic analytics charts to use the browser's timezone (#22660)
no issue

Originally I thought it would be smart to use the site's timezone for
all users, regardless of where they are located, so that the data will
always match if two people in different timezones are looking at the
same filters. However, this created some issues with the charts, since
they are all rendered in the browser's timezone, which led to some
inconsistencies.

Ultimately I still think it might make sense to use the site timezone,
but for now using the browser timezone feels reasonable and importantly
it actually works with the charts in the UI as you'd expect it to.