🐛 Fixed newsletter preferences for paid signups
ref https://linear.app/ghost/issue/ONC-803
ref https://github.com/TryGhost/Ghost/commit/b3a30d63ee

When members signup through the paid flow, we take the `newsletters`
payload and attach it to the Stripe checkout session as
metadata. Later this is passed to the `add` method of the Member
bookshelf model, and picked up by bookshelf-relations to set the newsletters.

When we stopped sending the `id` up in the linked commit, we broke
this funtionality because bookshelf-relations doesn't work on the
`name` property for Member Newsletters.