Fixed kpi pageview stats when filtering by pathname (#22674)
ref https://linear.app/ghost/issue/ANAL-162/

When filtering by path, the KPIs Pageviews metric is a bit odd. We
effectively need to introduce a new 'context' - in general, the KPIs
component shows the _session metrics_, but this is unintuitive when
filtering by path, because you don't care how many pageviews all the
sessions had that looked at the one particular path. Instead, you want
to see the # of pageviews *for that path only*.

We should also update this to handle time on page instead of the entire
session time in a separate change. Ideally, we end up reworking this
entirely and use a different MV for page-based metrics, and we'll need
to adjust the frontend to query a different endpoint.