Added tinybird-cli to devcontainer (#22662)
no issue

Currently we're using the bespoke tinybird-cli docker container to run
CLI commands. We have a command `yarn tb` to open a shell where the CLI
is accessible, but we can't run `tb` commands locally or leverage the
tinybird VSCode extension without installing the tinybird CLI.

This commit installs the Tinybird CLI in the development docker image,
so we can run `tb` commands in a shell in the main Ghost docker image.
It also introduces a `package.json` script to run an instance of the
Ghost container that just sleeps infinitely, so you can attach your IDE
to the container and run `tb` commands directly from your IDE's
terminal.

Not only is this more convenient than keeping an extra terminal tab open
to be able to run `tb` commands, it also lets us use the other tools in
the devcontainer (e.g. `jq`, `node`, etc) in Tinybird scripts, which we
couldn't do before because the tinybird-cli image is pretty barebones.