🐛 Fixed links being shown in incorrect order
ref https://linear.app/ghost/issue/ONC-845

Our API returns links ordered by click count descending. However we
merge together links with matching titles to aggregate the counts.

We didn't reorder based on the aggregated count, which resulted in a
broken looking UI, by adding a sort after the aggregation we display
the links in the correct order.