Fixed Tinybird CI jobs to use custom deploy script properly (#22725)
no issue

So far when we've made changes to Tinybird, we've only used the basic
`tb deploy` command. There is another option, to write a custom deploy
script, similar to a MySQL database migration, but currently the CI
pipeline doesn't work with this option.

This commit fixes the CI pipeline to use the custom deploy script for
the current version specified in the `.tinyenv` file if it exists, and
only if it doesn't exist to use the `tb deploy` command directly.

It also adds the `GHOST_UPSTREAM` environment variable as a pass-through
to the docker image, so that this value is available inside the
container — useful when attaching your IDE to the container and running
git commands like `yarn main`