Updated Tinybird `tbsql` function (#22728)
no issue

This commit enhances the `tbsql` utility function that we use to run SQL
queries against our Tinybird workspace during debugging.

- Enables `tbsql` in the Ghost container by using `$BASH_SOURCE` to
determine the path to the `.sql` files
- Adds a `--format` parameter, which can be set to `csv`, `json`, or
`human`. It defaults to `csv` for consistency
- Adds environment variable substitution, so we can use e.g.
`${TB_VERSION:-0}` in our SQL queries to target the current version
without having to find/replace

It also adds a few basic `.sql` queries to the repo which are useful
when working with our fixture tests.