Bumped Tinybird `.tinyenv` version to 0.0.2 (#22729)
no issue

- The Tinybird CI jobs aren't functioning right now, since there is a
custom deploy script for `0.0.1` and that is the version currently
specified in the `.tinyenv` file. Each "deploy" is just re-running the
same custom deploy script, and not actually deploying the changes in the
datafiles.
- The result is that the CI jobs are running tests against the current
implementation, without including the changes in the PR (because they
aren't deployed), which could lead to false passing results
- This bumps the version in the `.tinyenv` file, which doesn't do
anything other than fix the CI jobs, since there aren't any actual
changes to the datafiles in this commit