Fixed warnings in Tinybird test output for timezone tests (#22730)
no issue

- Our Tinybird test suite has sanity checks that compare the sum of
pageviews in a test result with the total number of rows in the fixture
data. This sanity check runs against all tests that start with `all_`.
- The timezone tests added last week used the `all_` prefix, but we
wouldn't expect all rows in the fixture to be represented, since some
rows are filtered out by timestamp
- This commit renames all the timezone tests so the sanity check won't
be run against these tests