Added route based app structure in ActivityPub (#22711)
ref AP-1006

Up until now the ActivityPub app has had a few main routes: inbox, feed,
profile, explore, notifications. This is fine for a starter, however it
limits the overall usability by not being able to refresh or refer to
(and eventually share) dedicated posts, profiles and other specific
pages in ActivityPub. This PR is restructuring the app with the
following details:

- Adds route to all detail views such as `inbox/:postId`, `feed/:postId`
and `profile/:handle`
- Completely removes the drawer in favor of using the middle area for
posts, replies and profiles
- Restructures components for better maintainability
- Refines related visual and usability details