Fixed stats page filtering by source (#22726)
ref https://linear.app/ghost/issue/ANAL-182
- moved all endpoints to use the same include file for filtering
sessions
- all endpoints used filtered sessions instead of directly querying hits
- re-enabled source filter on frontend
- added test coverage for dual filters + snapshots of expected output
from hits mv and the sessions pipe (which will be a view eventually)

Filtering by source was inconsistent across our various components,
which is why we disabled it. This set of changes unifies our filtering
logic across all endpoints and adds more coverage to raise a flag when
changes are disruptive.

---------

Co-authored-by: Chris Raible <chris@ghost.org>