Fixed `yarn docker:test:browser` command (#22745)
no issue

- Last week we added the `apps/stats` package, which wasn't accounted
for in the `Dockerfile` or the `compose.yml` file. This caused the admin
build to fail in the container due to a dependency mismatch, and the
`yarn docker:test:browser` command would also fail for the same reason.
- This commit updates the Docker setup to properly install the stats
app's dependencies, which fixes the admin build and the `yarn
docker:test:browser` command.