Updated Tinybird tests to run in parallel (#22754)
ref https://linear.app/ghost/issue/ANAL-189/run-tests-in-parallel

- As we've added more Tinybird tests to gain confidence in making
changes to Tinybird, our test suite has begun to become a bottleneck,
especially locally.
- Since all of our current tests are operating on the same static
fixture dataset, and we're just checking the result of API calls,
there's no reason these tests need to run in series, one after the
other.
- This change adds some logic to the exec_test.sh script to determine
the number of available CPU cores, and run the tests with {cpu_cores x
2} jobs.
- The exec_test.sh script also accepts a -j option, to override the
number of jobs — this allows us to run without parallelism by using
./scripts/exec_test.sh -j 1
- Before this commit, the CI test job was taking ~1 minute, 45 seconds
total, with the "Run fixture tests" step taking a full 1 minute. After,
the CI test job is taking < 1 minute, and the "Run fixture tests" step
is taking ~15 seconds.
- The difference when running locally is even more dramatic, at least in
west coast US — before, the exec_test script was taking ~1 minute 45
seconds, and now it's only taking ~6 seconds
- It also changes the default `yarn tb` command to use the main Ghost
image, so that we can more easily manage dependencies. The `exec_test`
script now depends on `parallel` being available, which isn't installed
in the tinybird-cli container. It is installed in the Ghost container,
along with the tinybird cli itself. For now, the previous version of the
`yarn tb` command has been moved to `yarn tbcli`, but I'll plan to
remove this once confirming the new `yarn tb` command works for everyone
on the team.