Updated Tinybird `gen_test_results` script to run in parallel (#22776)
no issue

- Currently generating test results using the `gen_test_results.sh`
script runs against each test in series, and as we've added more and
more tests, it's become a bottleneck when working on adding new tests.
- Following the same pattern used to parallelize the `exec_test.sh`
script, this commit modifies the `gen_test_results.sh` script to
generate the test results for all the tests in parallel, significantly
speeding up the operation.