Added Tinybird tests for the pathname filter for all endpoints (#22780)
ref
https://linear.app/ghost/issue/ANAL-194/automated-tests-for-all-queries-including-every-filter-plus-one-other

- This commit adds Tinybird fixture tests for filtering by pathname for
all endpoints that didn't previously have these tests

- This is part of an effort to generally improve our test coverage on
our Tinybird endpoints so that we can more confidently make changes to
them without breaking other endpoints or filter combinations

- The one excluded endpoint is the top_sources endpoint, because it is
giving incorrect results. I'll raise a separate PR with that test added
with the correct results so that we can fix the implementation to get it
to pass.