Updated mv_hits pipe to allow for non-url referrers (#22807)
no ref

Moved to coalesce to first try to keep the URL, and instead fall back to
the plain referrer field. This may include extra garbage entries, so
we'll open it up and we'll see what comes in. The intent is to include
`ghost-newsletter` which is traffic from our user newsletters.