Added basic linting for Tinybird datafiles (#22809)
no issue

- We've had issues where the TOKEN parameter is accidentally removed
from a pipe, which then requires manual intervention after a deployment
to fix the ugly red errors in the charts due to the "stats_page" token
not having permissions for the pipe.
- This commit adds a basic lint.sh script, which loops over all the
datafiles and runs some checks on them. Right now, there are only 3
simple checks, but we can add to this as we discover issues.

The checks included in this commit are:
- `check_token` — for any pipe starting with `api`, it checks that the
"stats_page" token is added to it
- `check_version` — for any datafile other than
`analytics_events.datasource`, it checks that there is a VERSION
parameter specified
- `check_consistent_version` — checks that all the versioned datafiles
are on the same version, to prevent accidentally shipping a partial
version.