Improved Tinybird scripts handling of TB_VERSION (#22814)
no issue

This commit includes various improvements to our Tinybird scripts, which
I made while working on improvements to the datafiles in a separate PR.
Pulling the tooling changes into this PR to separate the tooling changes
from the actual datafile changes.

- Adds a new `branch_deploy_append.sh` script which creates a new
branch, deploys local changes, and appends fixtures. This was a common
workflow for me that I was doing manually, so this script should save
some time.
- Added TB_VERSION to `.tinyenv` file, and exports the TB_VERSION from
this value in all scripts. Now we should only have to update this value
in this file to get all the scripts (i.e. `exec_test`,
`gen_test_results`, etc) to run against the current version of Tinybird.