Update CONTRIBUTING.md with Documentation guidance
When looking for how to propose updates to the [Install From Source Documentation](https://ghost.org/docs/install/source/), I found that the Docs were no longer open to public contributions. After some exploration, I found forum Posts from the Ghost Staff directing people to share feedback on Documentation via the forums.

It would have saved me some time searching to read this in the Contributing guide, and spell it out explicitly. The header mentions sharing "ideas" and asking questions in the forum, but I think updating the Docs is its own category, and people may be expecting them to be open source or live with the project code by default. 

This change is intended to make it easier for Contributors to know how to update the docs, and hopefully reduce frustration of needing to search for the right place to do it.  Thanks!