Updated hits view to handle non-url refs
ref https://linear.app/ghost/issue/PROD-196/

Updated the mv_hits pipe to parse values from the payload.referrer field that are not URLs. This was tricky because the JSON data type handling seems incomplete. We saw null values that would persist and not be detectable by standards functions like IS NULL or = 'null', etc. By checking at the time of parsing the JSON, it seems to propagate normally.

As a final note, the url fns in Tinybird return a value in any case, so they do not work with coalesce without double calculating, which is a bit annoying.