Added Tinybird linting check to pre-commit hook (#22833)
ref
https://linear.app/ghost/issue/PROD-171/formatting-linting-the-pipe-and-datasource-files

- We now have a linting job that will catch any issues with the
`VERSION` or `TOKEN` parameter in our Tinybird datafiles, which will run
in CI and prevent merging if either of these parameters is missing or
incorrect.
- This commit adds the same check as a pre-commit hook to shorten the
feedback loop, so we don't have to wait until CI runs or run the linter
manually to notice the error.
- The linting job should only run in the pre-commit hook if a
`*.datasource` or `*.pipe` file in `ghost/web-analytics` is modified.