Added `wa` wrapper CLI for the Tinybird CLI (#22841)
ref
https://linear.app/ghost/issue/PROD-208/wrap-our-tinybird-scripts-in-a-basic-cli

- We currently have a handful of bash scripts which we use many times
daily when working on Tinybird, and we're invoking the scripts using
`./scripts/{script_name}.sh`. Major "first world problem" here, but it
would be nice to avoid typing so many characters multiple times every
day.
- This commit introduces a tiny bash wrapper called `wa` which then
delegates to the existing script files. The CLI is available using the
`yarn tb` script to launch a shell, then you can run e.g. `wa branch` to
create a new Tinybird branch with a randomly generated name.
- More details on the usage are in the README.md, and you can also run
`wa help` to output all the available commands and options
- This is just an experimental starting point that we can keep building
on if it feels right, or we can delete it later if we find it's not
useful.