Ignored Tinybird CLI version warning (#22843)
no issue

- The Tinybird CLI outputs a warning message when the version of the CLI
is not the latest version.
- This can be a problem because the warning message is included in the
test results files, which causes tests to fail.
- This is particularly annoying because we are on the latest version of
the CLI, but we are getting the warning for development builds of the
CLI, not just the latest stable release.
- This commit adds the `TB_VERSION_WARNING` environment variable to the
shell opened with `yarn tb` and to the Github Actions workflows to
suppress the warning message.