Added unit tests for `ghost-stats` script (#22842)
ref https://linear.app/ghost/issue/PROD-195
- added unit tests for `ghost-stats`

We don't have a great model for testing scripts in the `/public`
directory. Ideally we'd build these and be able to more easily unit test
the pieces. Since this may change a fair amount as we flesh out this
functionality, this simply loads the script using `JSDom` and mocks a
lot of the browser capabilities. We could add some e2e tests later on.

The main thing to watch here is how the data is formatted and how we
handle pieces like the referrer.