Migrated to web analytics build to TB Forward (#22859)
ref https://linear.app/ghost/issue/PROD-214
- migrated all Tinybird assets to a new directory structure matching
Tinybird Forward
- migrated all tests to the new `.yaml` structure
- updated scripts to handle Tinybird local + new directory
- updated config handling to allow for TB Local config (+ some general
improvements/cleaning up)

Tinybird Forward is the miles ahead of TB Classic when in comes to DevX
and is going to be the future of their product. With that in mind, it'd
be best to move over to it while we're building out the web analytics
functionality. For the moment, _we are running the two builds in
parallel_ while we figure out deployments (CI/CD) with Tinybird Forward,
as the `VERSION` field is fully deprecated.

The primary
[benefit](https://www.tinybird.co/docs/forward/get-started/concepts) is
being able to develop locally with Tinybird Local, which runs in a
Docker container. This commit adds the support to run alongside/within
Ghost. See the readme file as part of this commit for instructions on
how to configure Ghost's config.