Added tinybird conversion script to convert csv to ndjson (#22892)
ref https://linear.app/ghost/issue/PROD-224/

This is a bit fussy because the JSON field is imperfectly supported and
it's not noted where it fails. Added via comments, the intended use for
this script is:
- export csv file from Tinybird Classic workspace, likely using the CLI
to get an un-pagingated output
- execute script on csv file
- output is a ndjson file that can be imported using the TB Forward CLI,
e.g. `tb datasource append analytics_events ~/dev/output.ndjson`

This should allow us to migrate data from TB Classic to TB Forward. I
don't expect their S3 bucket workflow to actually work for us, and all
the tooling fails when trying to write sql to csv or otherwise pipe the
output the csv/json file.