Increased max usage count of magic links from 3 -> 7
ref https://linear.app/ghost/issue/PROD-227

Inbox security clients will make requests to URLs found in emails in an attempt
to detect security threats. This causes a problem when magic links or tokens
are single use. Because they are invalidated by the time the user is presented
with them. We originally had a usage count set to 3 to alleviate this problem,
but we are still seeing issues with clients like Barracuda. As a short term
experiment we are raising the max usage from 3 -> 7 and monitoring the impact.