Added `trafficAnalytics` beta flag (#22974)
ref
https://linear.app/ghost/issue/PROD-635/enable-trafficanalytics-with-a-private-beta-flag-instead-of-via-config

The Traffic Analytics private beta is currently enabled via config. This
is a pain, because it requires a code change in Zuul to enable the beta
for new sites.

Putting the Traffic Analytics features behind a private beta flag makes
it easier to enable the feature without having to make a code/config
change. This commit just adds the beta flag on the backend, but doesn't
do anything with it yet.